Automation_BackUP_2/Automation/requirements.txt
selenium==4.16.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.1.0
certifi==2023.11.17
idna==3.6
charset-normalizer==3.3.2
webdriver-manager==4.0.1
pdf2image==1.17
Pillow==9.5.0
